requirements.txt
alembic==0.9.6
apturl==0.5.2
asn1crypto==0.24.0
Babel==2.5.1
blinker==1.4
Brlapi==0.6.6
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info==0.18
dominate==2.3.1
elasticsearch==6.3.1
flake8==3.6.0
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
isodate==0.6.0
itsdangerous==0.24
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pycairo==1.16.2
pycodestyle==2.4.0
pycrypto==2.6.1
pycups==1.9.73
pyenchant==2.0.0
pyflakes==2.0.0
pygobject==3.26.1
PyJWT==1.5.3
pyliblo==0.10.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
python-apt==1.6.2
python-dateutil==2.6.1
python-debian==0.1.32
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2017.2
pyxdg==0.25
PyYAML==3.12
rdflib==4.2.1
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
SPARQLWrapper==1.7.6
SQLAlchemy==1.1.14
system-service==0.3
systemd-python==234
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.3
visitor==0.1.3
wadllib==1.3.2
webencodings==0.5
Werkzeug==0.14.1
WTForms==2.1
xkit==0.0.0
zope.interface==4.3.2
